Moved common function to util file.
